+++ Plugins/Reqnroll.TUnit.ReqnrollPlugin/TUnitRuntimeProvider.cs
@@ -1,4 +1,6 @@
-﻿using Reqnroll.UnitTestProvider;
+﻿using System;
+using Reqnroll.ErrorHandling;
+using Reqnroll.UnitTestProvider;
 using TUnit.Core.Exceptions;
 
 namespace Reqnroll.TUnit.ReqnrollPlugin;
@@ -7,7 +9,7 @@ public class TUnitRuntimeProvider : IUnitTestRuntimeProvider
 {
     public void TestPending(string message)
     {
-        throw new PendingStepException(message);
+        throw new PendingScenarioException(message);
     }
 
     public void TestInconclusive(string message)
@@ -19,4 +21,10 @@ public class TUnitRuntimeProvider : IUnitTestRuntimeProvider
     {
         Skip.Test(message);
     }
+
+    public ScenarioExecutionStatus? DetectExecutionStatus(Exception exception) => exception switch
+    {
+        InconclusiveTestException or SkipTestException => ScenarioExecutionStatus.Skipped,
+        _ => null
+    };
 }