+++ Tests/Reqnroll.RuntimeTests/Infrastructure/TestThreadExecutionEventPublisherTests.cs
@@ -22,9 +22,9 @@ namespace Reqnroll.RuntimeTests.Infrastructure
             await testExecutionEngine.StepAsync(StepDefinitionKeyword.Given, null, "foo", null, null);
             
             _testThreadExecutionEventPublisher.Verify(te =>
-                    te.PublishEventAsync(It.Is<StepStartedEvent>(e => e.ScenarioContext.Equals(scenarioContext) &&
-                                                                 e.FeatureContext.Equals(featureContainer.Resolve<FeatureContext>()) &&
-                                                                 e.StepContext.Equals(contextManagerStub.Object.StepContext))), 
+                    te.PublishEventAsync(It.Is<StepStartedEvent>(e => e.ScenarioContext.Equals(_scenarioContext) &&
+                                                                 e.FeatureContext.Equals(_featureContainer.Resolve<FeatureContext>()) &&
+                                                                 e.StepContext.Equals(_contextManagerStub.Object.StepContext))), 
                                                       Times.Once);
         }
 
@@ -47,8 +47,8 @@ namespace Reqnroll.RuntimeTests.Infrastructure
         {
             var stepDef = RegisterStepDefinition().Object;
             TimeSpan expectedDuration = TimeSpan.FromSeconds(5);
-            methodBindingInvokerMock
-                .Setup(i => i.InvokeBindingAsync(stepDef, contextManagerStub.Object, It.IsAny<object[]>(), testTracerStub.Object, It.IsAny<DurationHolder>()))
+            _methodBindingInvokerMock
+                .Setup(i => i.InvokeBindingAsync(stepDef, _contextManagerStub.Object, It.IsAny<object[]>(), _testTracerStub.Object, It.IsAny<DurationHolder>()))
                 .Callback((IBinding _, IContextManager _, object[] arguments, ITestTracer _, DurationHolder durationHolder) => durationHolder.Duration = expectedDuration)
                 .ReturnsAsync(new object());
             var testExecutionEngine = CreateTestExecutionEngine();
@@ -73,9 +73,9 @@ namespace Reqnroll.RuntimeTests.Infrastructure
             
             _testThreadExecutionEventPublisher.Verify(te =>
                 te.PublishEventAsync(It.Is<StepFinishedEvent>(e => 
-                                                             e.ScenarioContext.Equals(scenarioContext) &&
-                                                             e.FeatureContext.Equals(featureContainer.Resolve<FeatureContext>()) &&
-                                                             e.StepContext.Equals(contextManagerStub.Object.StepContext))),
+                                                             e.ScenarioContext.Equals(_scenarioContext) &&
+                                                             e.FeatureContext.Equals(_featureContainer.Resolve<FeatureContext>()) &&
+                                                             e.StepContext.Equals(_contextManagerStub.Object.StepContext))),
                                                       Times.Once);
         }
         
@@ -85,7 +85,7 @@ namespace Reqnroll.RuntimeTests.Infrastructure
             RegisterStepDefinition();
             var testExecutionEngine = CreateTestExecutionEngine();
             //a step will be skipped if the ScenarioExecutionStatus is not OK
-            scenarioContext.ScenarioExecutionStatus = ScenarioExecutionStatus.TestError;
+            _scenarioContext.ScenarioExecutionStatus = ScenarioExecutionStatus.TestError;
 
             await testExecutionEngine.StepAsync(StepDefinitionKeyword.Given, null, "foo", null, null);
             
@@ -99,13 +99,13 @@ namespace Reqnroll.RuntimeTests.Infrastructure
             var hookType = HookType.AfterScenario;
             TimeSpan expectedDuration = TimeSpan.FromSeconds(5);
             var expectedHookBinding = new HookBinding(new Mock<IBindingMethod>().Object, hookType, null, 1);
-            methodBindingInvokerMock
-                .Setup(i => i.InvokeBindingAsync(expectedHookBinding, contextManagerStub.Object, It.IsAny<object[]>(), testTracerStub.Object, It.IsAny<DurationHolder>()))
+            _methodBindingInvokerMock
+                .Setup(i => i.InvokeBindingAsync(expectedHookBinding, _contextManagerStub.Object, It.IsAny<object[]>(), _testTracerStub.Object, It.IsAny<DurationHolder>()))
                 .Callback((IBinding _, IContextManager _, object[] arguments, ITestTracer _, DurationHolder durationHolder) => durationHolder.Duration = expectedDuration)
                 .ReturnsAsync(new object());
             var testExecutionEngine = CreateTestExecutionEngine();
 
-            await testExecutionEngine.InvokeHookAsync(methodBindingInvokerMock.Object, expectedHookBinding, hookType);
+            await testExecutionEngine.InvokeHookAsync(_methodBindingInvokerMock.Object, expectedHookBinding, hookType);
 
             _testThreadExecutionEventPublisher.Verify(te =>
                 te.PublishEventAsync(It.Is<HookBindingStartedEvent>(e =>
@@ -123,13 +123,13 @@ namespace Reqnroll.RuntimeTests.Infrastructure
         {
             var testExecutionEngine = CreateTestExecutionEngine();
 
-            testExecutionEngine.OnScenarioInitialize(scenarioInfo, ruleInfo);
+            testExecutionEngine.OnScenarioInitialize(_scenarioInfo, _ruleInfo);
             await testExecutionEngine.OnScenarioStartAsync();
 
             _testThreadExecutionEventPublisher.Verify(te =>
                 te.PublishEventAsync(It.Is<ScenarioStartedEvent>(e =>
-                                                                e.ScenarioContext.Equals(scenarioContext) &&
-                                                                e.FeatureContext.Equals(featureContainer.Resolve<FeatureContext>()))),
+                                                                e.ScenarioContext.Equals(_scenarioContext) &&
+                                                                e.FeatureContext.Equals(_featureContainer.Resolve<FeatureContext>()))),
                                                       Times.Once);
         }
 
@@ -138,15 +138,15 @@ namespace Reqnroll.RuntimeTests.Infrastructure
         {
             var testExecutionEngine = CreateTestExecutionEngine();
 
-            testExecutionEngine.OnScenarioInitialize(scenarioInfo, ruleInfo);
+            testExecutionEngine.OnScenarioInitialize(_scenarioInfo, _ruleInfo);
             await testExecutionEngine.OnScenarioStartAsync();
             await testExecutionEngine.OnAfterLastStepAsync();
             await testExecutionEngine.OnScenarioEndAsync();
             
             _testThreadExecutionEventPublisher.Verify(te =>
                 te.PublishEventAsync(It.Is<ScenarioFinishedEvent>(e =>
-                                                                 e.ScenarioContext.Equals(scenarioContext) &&
-                                                                 e.FeatureContext.Equals(featureContainer.Resolve<FeatureContext>()))),
+                                                                 e.ScenarioContext.Equals(_scenarioContext) &&
+                                                                 e.FeatureContext.Equals(_featureContainer.Resolve<FeatureContext>()))),
                                                       Times.Once);
         }
 
@@ -155,11 +155,11 @@ namespace Reqnroll.RuntimeTests.Infrastructure
         {
             var testExecutionEngine = CreateTestExecutionEngine();
 
-            var hookMock = CreateHookMock(afterScenarioEvents);
-            methodBindingInvokerMock.Setup(i => i.InvokeBindingAsync(hookMock.Object, contextManagerStub.Object, null, testTracerStub.Object, It.IsAny<DurationHolder>()))
+            var hookMock = CreateHookMock(_afterScenarioEvents);
+            _methodBindingInvokerMock.Setup(i => i.InvokeBindingAsync(hookMock.Object, _contextManagerStub.Object, null, _testTracerStub.Object, It.IsAny<DurationHolder>()))
                                     .Throws(new Exception("simulated hook error"));
 
-            testExecutionEngine.OnScenarioInitialize(scenarioInfo, ruleInfo);
+            testExecutionEngine.OnScenarioInitialize(_scenarioInfo, _ruleInfo);
             await testExecutionEngine.OnScenarioStartAsync();
             await testExecutionEngine.OnAfterLastStepAsync();
             await FluentActions.Awaiting(testExecutionEngine.OnScenarioEndAsync)
@@ -167,8 +167,8 @@ namespace Reqnroll.RuntimeTests.Infrastructure
             
             _testThreadExecutionEventPublisher.Verify(te =>
                 te.PublishEventAsync(It.Is<ScenarioFinishedEvent>(e =>
-                                                                 e.ScenarioContext.Equals(scenarioContext) &&
-                                                                 e.FeatureContext.Equals(featureContainer.Resolve<FeatureContext>()))),
+                                                                 e.ScenarioContext.Equals(_scenarioContext) &&
+                                                                 e.FeatureContext.Equals(_featureContainer.Resolve<FeatureContext>()))),
                                                       Times.Once);
         }
 
@@ -179,9 +179,9 @@ namespace Reqnroll.RuntimeTests.Infrastructure
             RegisterStepDefinition();
 
             await testExecutionEngine.OnTestRunStartAsync();
-            await testExecutionEngine.OnFeatureStartAsync(featureInfo);
+            await testExecutionEngine.OnFeatureStartAsync(_featureInfo);
 
-            testExecutionEngine.OnScenarioInitialize(scenarioInfo, ruleInfo);
+            testExecutionEngine.OnScenarioInitialize(_scenarioInfo, _ruleInfo);
             await testExecutionEngine.OnScenarioStartAsync();
             await testExecutionEngine.StepAsync(StepDefinitionKeyword.Given, null, "foo", null, null);
             await testExecutionEngine.OnAfterLastStepAsync();
@@ -205,22 +205,22 @@ namespace Reqnroll.RuntimeTests.Infrastructure
         {
             var testExecutionEngine = CreateTestExecutionEngine();
 
-            testExecutionEngine.OnScenarioInitialize(scenarioInfo, ruleInfo);
+            testExecutionEngine.OnScenarioInitialize(_scenarioInfo, _ruleInfo);
             await testExecutionEngine.OnScenarioSkippedAsync();
             await testExecutionEngine.OnAfterLastStepAsync();
             await testExecutionEngine.OnScenarioEndAsync();
             
             _testThreadExecutionEventPublisher.Verify(te =>
                 te.PublishEventAsync(It.Is<ScenarioStartedEvent>(e =>
-                                                                e.ScenarioContext.Equals(scenarioContext) &&
-                                                                e.FeatureContext.Equals(featureContainer.Resolve<FeatureContext>()))),
+                                                                e.ScenarioContext.Equals(_scenarioContext) &&
+                                                                e.FeatureContext.Equals(_featureContainer.Resolve<FeatureContext>()))),
                                                       Times.Once);
             _testThreadExecutionEventPublisher.Verify(te =>
                 te.PublishEventAsync(It.IsAny<ScenarioSkippedEvent>()), Times.Once);
             _testThreadExecutionEventPublisher.Verify(te =>
                 te.PublishEventAsync(It.Is<ScenarioFinishedEvent>(e =>
-                                                                 e.ScenarioContext.Equals(scenarioContext) &&
-                                                                 e.FeatureContext.Equals(featureContainer.Resolve<FeatureContext>()))),
+                                                                 e.ScenarioContext.Equals(_scenarioContext) &&
+                                                                 e.FeatureContext.Equals(_featureContainer.Resolve<FeatureContext>()))),
                                                       Times.Once);
         }
 
@@ -229,11 +229,11 @@ namespace Reqnroll.RuntimeTests.Infrastructure
         {
             var testExecutionEngine = CreateTestExecutionEngine();
 
-            await testExecutionEngine.OnFeatureStartAsync(featureInfo);
+            await testExecutionEngine.OnFeatureStartAsync(_featureInfo);
             
             _testThreadExecutionEventPublisher.Verify(te =>
                 te.PublishEventAsync(It.Is<FeatureStartedEvent>(e =>
-                                                               e.FeatureContext.Equals(featureContainer.Resolve<FeatureContext>()))),
+                                                               e.FeatureContext.Equals(_featureContainer.Resolve<FeatureContext>()))),
                                                       Times.Once);
         }
 
@@ -246,7 +246,7 @@ namespace Reqnroll.RuntimeTests.Infrastructure
             
             _testThreadExecutionEventPublisher.Verify(te =>
                 te.PublishEventAsync(It.Is<FeatureFinishedEvent>(e => 
-                                                                e.FeatureContext.Equals(featureContainer.Resolve<FeatureContext>()))),
+                                                                e.FeatureContext.Equals(_featureContainer.Resolve<FeatureContext>()))),
                                                       Times.Once);
         }
 
@@ -255,8 +255,8 @@ namespace Reqnroll.RuntimeTests.Infrastructure
         {
             var testExecutionEngine = CreateTestExecutionEngine();
 
-            var hookMock = CreateHookMock(afterFeatureEvents);
-            methodBindingInvokerMock.Setup(i => i.InvokeBindingAsync(hookMock.Object, contextManagerStub.Object, null, testTracerStub.Object, It.IsAny<DurationHolder>()))
+            var hookMock = CreateHookMock(_afterFeatureEvents);
+            _methodBindingInvokerMock.Setup(i => i.InvokeBindingAsync(hookMock.Object, _contextManagerStub.Object, null, _testTracerStub.Object, It.IsAny<DurationHolder>()))
                                     .Throws(new Exception("simulated hook error"));
 
             await FluentActions.Awaiting(testExecutionEngine.OnFeatureEndAsync)
@@ -264,7 +264,7 @@ namespace Reqnroll.RuntimeTests.Infrastructure
             
             _testThreadExecutionEventPublisher.Verify(te =>
                 te.PublishEventAsync(It.Is<FeatureFinishedEvent>(e => 
-                                                                e.FeatureContext.Equals(featureContainer.Resolve<FeatureContext>()))),
+                                                                e.FeatureContext.Equals(_featureContainer.Resolve<FeatureContext>()))),
                                                       Times.Once);
         }
 
@@ -295,8 +295,8 @@ namespace Reqnroll.RuntimeTests.Infrastructure
         {
             var testExecutionEngine = CreateTestExecutionEngine();
 
-            var hookMock = CreateHookMock(afterTestRunEvents);
-            methodBindingInvokerMock.Setup(i => i.InvokeBindingAsync(hookMock.Object, contextManagerStub.Object, null, testTracerStub.Object, It.IsAny<DurationHolder>()))
+            var hookMock = CreateHookMock(_afterTestRunEvents);
+            _methodBindingInvokerMock.Setup(i => i.InvokeBindingAsync(hookMock.Object, _contextManagerStub.Object, null, _testTracerStub.Object, It.IsAny<DurationHolder>()))
                                     .Throws(new Exception("simulated hook error"));
 
             await FluentActions.Awaiting(testExecutionEngine.OnTestRunEndAsync)